+++ src/main/java/org/truffleruby/core/monitor/package-info.java
@@ -1,4 +1,4 @@
-@SuppressPackageWarnings({ "truffle-sharing", "truffle-inlining" })
+@SuppressPackageWarnings({ "truffle-inlining" })
 package org.truffleruby.core.monitor;
 
 import com.oracle.truffle.api.dsl.SuppressPackageWarnings;